refactor: move FormatT into response_formatting
