+++ python/mirascope/llm/agents/async_agent.py
@@ -5,9 +5,10 @@ from dataclasses import dataclass
 
 from ..content import UserContent
 from ..context import Context
+from ..response_formatting import FormatT
 from ..responses import Response
 from ..streams import AsyncStream
-from ..types import DepsT, FormatT
+from ..types import DepsT
 from .base_agent import BaseAgent
