+++ python/mirascope/llm/clients/base.py
@@ -9,13 +9,14 @@ from typing import Any, Generic, TypedDict
 from typing_extensions import TypeVar
 
 from ..context import Context
+from ..response_formatting import FormatT
 from ..responses import Response
 from ..streams import (
     AsyncStream,
     Stream,
 )
 from ..tools import ContextToolDef, ToolDef
-from ..types import DepsT, FormatT
+from ..types import DepsT
 from .register import REGISTERED_LLMS
 
 ProviderMessageT = TypeVar("ProviderMessageT")
