+++ python/mirascope/llm/models/base.py
@@ -18,7 +18,8 @@ if TYPE_CHECKING:
     from ..clients import REGISTERED_LLMS
 
 
-from ..types import DepsT, FormatT
+from ..response_formatting import FormatT
+from ..types import DepsT
 
 
 class LLM(Generic[ProviderMessageT, ParamsT, ClientT]):
