+++ python/mirascope/llm/response_formatting/response_format.py
@@ -3,7 +3,18 @@
 from dataclasses import dataclass
 from typing import Any, Generic, Literal, Protocol, runtime_checkable
 
-from ..types import FormatCovariantT, FormatT
+from typing_extensions import TypeVar
+
+from ..types import FormatCovariantT
+
+FormatT = TypeVar("FormatT", bound=object | None, default=None)
+"""Type variable for structured response format types.
+
+This TypeVar represents the type of structured output format that LLM responses
+can be parsed into, such as Pydantic models, dataclasses, or custom classes.
+It can be None for unstructured responses and defaults to None when no specific
+format is required.
+"""
 
 
 class JsonParserFn(Protocol[FormatCovariantT]):
