+++ python/mirascope/llm/types/__init__.py
@@ -7,7 +7,6 @@ from .type_vars import (
     ChunkT,
     DepsT,
     FormatCovariantT,
-    FormatT,
     P,
     ToolCovariantT,
     ToolReturnT,
@@ -19,7 +18,6 @@ __all__ = [
     "Dataclass",
     "DepsT",
     "FormatCovariantT",
-    "FormatT",
     "Jsonable",
     "P",
     "ToolCovariantT",
